refactor: [#203] replace old e2e_config_tests with black-box version
- Delete old white-box e2e_config_tests.rs
- Rename e2e_config_tests_new.rs to e2e_config_tests.rs
- Update environment name from e2e-config-new to e2e-config
- Remove obsolete e2e-config-tests-new binary entry from Cargo.toml
- Update doc comments and example references

This ensures the black-box E2E test runs in the GitHub Actions
workflow (.github/workflows/test-e2e-config.yml).
